requirements.txt
esptool
rshell>=0.0.30,<1.0.0